requirements.txt
streamlit>=1.40.1
pyaudio>=0.2.14
openai-whisper>=20240930
crewai>=0.80.0
git+https://github.com/SWivid/F5-TTS.git
python-json-logger>=2.0.7
fire>=0.7.0
pytest>=8.0.0
black>=24.1.1
pylint>=3.0.3
SoundCard>=0.4.3
PyAudio>=0.2.14
numpy>=1.24.0
pyperclip>=1.9.0
sounddevice>=0.5.1
fire>=0.4.0
keyboard>=0.13.5 